[ARQ-600] Improve Tomcat codebase - Managed  * Add new tomcat-managed-common module  * Refactor common "managed" code into it.  * Use IOUtilDelegator to read server responses in unit tests.
