+++ pom.xml
@@ -123,8 +123,8 @@
             <build>
                 <plugins>
                     <!--
-                        surefire default forks the JVM for testing, so we need to pass on the maven.repo.local variable so the MavenResovler used in
-                        the test cases can read from the same lcoal repo.
+                        surefire default forks the JVM for testing, so we need to pass on the maven.repo.local variable so the MavenResolver used in
+                        the test cases can read from the same local repo.
                      -->
                     <plugin>
                         <artifactId>maven-surefire-plugin</artifactId>
@@ -140,6 +140,7 @@
     </profiles>
     <modules>
         <module>tomcat-common</module>
+        <module>tomcat-managed-common</module>
         <module>tomcat-managed-5.5</module>
         <module>tomcat-embedded-6</module>
         <module>tomcat-remote-6</module>
