+++ tomcat-managed-5.5/pom.xml
@@ -21,6 +21,17 @@
             <artifactId>arquillian-tomcat-common</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.jboss.arquillian.container</groupId>
+            <artifactId>arquillian-tomcat-managed-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.arquillian.container</groupId>
+            <artifactId>arquillian-tomcat-managed-common</artifactId>
+            <type>test-jar</type>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- org.jboss.arquillian -->
         <dependency>
