+++ tomcat-managed-5.5/src/test/java/org/jboss/arquillian/container/tomcat/managed_5_5/TomcatManagedClientTestCase.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat Middleware LLC, and individual contributors
+ * Copyright 2014, Red Hat Middleware LLC, and individual contributors
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -21,9 +21,11 @@ import java.net.URL;
 import java.util.logging.Logger;
 
 import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.container.tomcat.managed.MyServlet;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.IOUtilDelegator;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.shrinkwrap.descriptor.api.Descriptors;
@@ -82,7 +84,7 @@ public class TomcatManagedClientTestCase
    // -------------------------------------------------------------------------------------||
 
    /**
-    * Ensures the {@link HelloWorldServlet} returns the expected response
+    * Ensures the {@link MyServlet} returns the expected response
     */
    @Test
    public void shouldBeAbleToInvokeServletInDeployedWebApp(@ArquillianResource URL contextRoot) throws Exception
@@ -93,11 +95,8 @@ public class TomcatManagedClientTestCase
       URL url = new URL(contextRoot, "Test");
       InputStream in = url.openConnection().getInputStream();
 
-      byte[] buffer = new byte[10000];
-      int len = in.read(buffer);
-      String httpResponse = "";
-      for (int q = 0; q < len; q++)
-         httpResponse += (char) buffer[q];
+       byte[] buffer = IOUtilDelegator.asByteArray(in);
+       String httpResponse = new String(buffer);
 
       // Test
       Assert.assertEquals("Expected output was not equal by value", expected, httpResponse);
