+++ tomcat-managed-5.5/src/test/java/org/jboss/arquillian/container/tomcat/managed_5_5/TomcatManagedInContainerTestCase.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat Middleware LLC, and individual contributors
+ * Copyright 2014, Red Hat Middleware LLC, and individual contributors
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -19,14 +19,15 @@ package org.jboss.arquillian.container.tomcat.managed_5_5;
 import java.io.InputStream;
 import java.net.URL;
 import java.util.logging.Logger;
-
 import javax.annotation.Resource;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
+import org.jboss.arquillian.container.tomcat.managed.MyServlet;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.IOUtilDelegator;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Assert;
 import org.junit.Test;
@@ -85,13 +86,8 @@ public class TomcatManagedInContainerTestCase
       URL url = new URL(contextRoot, "Test");
       InputStream in = url.openConnection().getInputStream();
 
-      byte[] buffer = new byte[10000];
-      int len = in.read(buffer);
-      String httpResponse = "";
-      for (int q = 0; q < len; q++)
-      {
-         httpResponse += (char) buffer[q];
-      }
+       byte[] buffer = IOUtilDelegator.asByteArray(in);
+       String httpResponse = new String(buffer);
 
       // Test
       Assert.assertEquals("Expected output was not equal by value", expected, httpResponse);
