+++ tomcat-managed-6/src/test/resources/in-container-web.xml
@@ -16,7 +16,7 @@
 
     <servlet>
         <servlet-name>TestServlet</servlet-name>
-        <servlet-class>org.jboss.arquillian.container.tomcat.managed_6.MyServlet</servlet-class>
+        <servlet-class>org.jboss.arquillian.container.tomcat.managed.MyServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
