+++ tomcat-managed-common/src/main/java/org/jboss/arquillian/container/tomcat/managed/ShrinkWrapUtil.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.arquillian.container.tomcat.managed_7;
+package org.jboss.arquillian.container.tomcat.managed;
 
 import java.io.File;
 import java.io.FileOutputStream;
